requirements.txt
influxdb
bottom
uvloop
jinja2
boto3
aiohttp
aioamqp
PyYAML
